app/requirements.txt
clickhouse-driver==0.2.8
Faker==26.0.0
iniconfig==2.0.0
kafka==1.3.5
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzlocal==5.2
